+++ PDEs/Math/Discretization/discretization.h
@@ -1,10 +1,11 @@
 #ifndef SPATIAL_DISCRETIZATION_H
 #define SPATIAL_DISCRETIZATION_H
 
-#include "Grid/grid_structs.h"
-#include "Grid/mesh.h"
+#include "grid_structs.h"
+#include "mesh.h"
 
-#include <memory>
+#include <cinttypes>
+
 
 namespace math
 {
@@ -43,16 +44,16 @@ public:
 
 public:
   /// Get the number of nodes in the discretization.
-  virtual size_t n_nodes() const { return 0; }
+  virtual uint64_t n_nodes() const { return 0; }
 
   /// Get the number of DoFs in the spatial discretization.
-  virtual size_t n_dofs(const size_t n_components) const { return 0; }
+  virtual uint64_t n_dofs(const uint64_t n_components) const { return 0; }
 
   /// Get the number of nodes per cell.
-  virtual size_t nodes_per_cell() const { return 0; }
+  virtual uint64_t nodes_per_cell() const { return 0; }
 
   /// Get the number of DoFs per cell.
-  virtual size_t dofs_per_cell(const size_t n_components) const { return 0; }
+  virtual uint64_t dofs_per_cell(const uint64_t n_components) const { return 0; }
 
    /// Get the location of the nodes on a cell.
   virtual std::vector<grid::Point> nodes(const grid::Cell& cell) const = 0;
@@ -62,7 +63,7 @@ public:
    *
    * This routine defines the column indices of non-zero entries per row for a
    * problem with the specified number of components. If the \p is_coupled flag
-   * is set to \p true, it is assumed that all components are coupled to one
+   * is insert to \p true, it is assumed that all components are coupled to one
    * another, otherwise, it is assumed that the system is uncoupled in all
    * components.
    *
@@ -72,8 +73,8 @@ public:
    *   solution components.
    */
   virtual void
-  make_sparsity_pattern(std::vector<std::vector<size_t>> pattern,
-                        const size_t n_components = 1,
+  make_sparsity_pattern(std::vector<std::vector<uint64_t>> pattern,
+                        const uint64_t n_components = 1,
                         const bool is_coupled = false) const = 0;
 };
